Added script descriptions and changed names
^
